stan na 5.12.2016 17:15

+++ HelloNeuralNetwork/Models/DigitImageForEvaluating.cs
@@ -3,7 +3,7 @@
     public class DigitImageForEvaluating : DigitImage
     {
 
-        public int Label { get; }
+        public byte Label { get; }
 
         public DigitImageForEvaluating(double[,] pixels, byte label) : base(pixels)
         {
